// src/model/IErrorRanelagh.ts
export interface IErrorRanelagh {
    refDocumento: string;
    responsable: string;
    detectadoPor: string;
    puestoResponsable: string;
    sectorResponsable: string;
    comentarioError: string;
    fechaRegistro: string;
    fechaResolucion: string;
    emitidoPor: string;
    id?: number | null;
    comparado?: boolean;
}